+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/util/LocalUISettings.java
@@ -1,7 +1,7 @@
 package de.mephisto.vpin.ui.util;
 
 import de.mephisto.vpin.commons.utils.PropertiesStore;
-import de.mephisto.vpin.ui.tournaments.dialogs.TournamentEditDialogController;
+import edu.umd.cs.findbugs.annotations.Nullable;
 import javafx.scene.shape.Rectangle;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,12 +11,31 @@ import java.io.File;
 public class LocalUISettings {
   private final static Logger LOG = LoggerFactory.getLogger(LocalUISettings.class);
 
+  public static final String LAST_FOLDER_SELECTION = "lastFolderSelection";
+
   private static PropertiesStore store;
 
   static {
     File propertiesFile = new File("config/settings.properties");
     propertiesFile.getParentFile().mkdirs();
     store = PropertiesStore.create(propertiesFile);
+  }
+
+  public static void saveLastFolderLocation(@Nullable File file) {
+    if (file != null) {
+      if (file.isFile()) {
+        file = file.getParentFile();
+      }
+      store.set(LAST_FOLDER_SELECTION, file.getAbsolutePath());
+    }
+  }
+
+  @Nullable
+  public static File getLastFolderSelection() {
+    if (store.containsKey(LAST_FOLDER_SELECTION)) {
+      return new File(store.get(LAST_FOLDER_SELECTION));
+    }
+    return null;
   }
 
   public static void saveLocation(int x, int y, int width, int height) {